Add Movies!
That's right, we can now add movies to the chosen Genre! We can also NOT
delete the Genre if there are movies existing within the Genre itself.
Now I'm just missing Edit and Delete movies, and we're done!!

+++ ArmandoGarcin_MovieDB/TheMovieDB/TheMovieDB/Controllers/MovieLibController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Data.Entity;
 using System.Web;
 using System.Web.Mvc;
 using TheMovieDB.MovieData;
@@ -42,28 +43,39 @@ namespace TheMovieDB.Controllers
         [AllowAnonymous]
         public ActionResult AddGenre(AddGenreModel _model)
         {
-            int tID = GetGenreID(_model.GenreName);
-
-            IdentityDBContext db = new IdentityDBContext();
-
-            foreach(Genre genre in db.Genres)
+            if(ModelState.IsValid)
             {
-                if(genre.GenreID == tID)
+                int tID = GetGenreID(_model.GenreName);
+
+                IdentityDBContext db = new IdentityDBContext();
+
+                foreach (Genre genre in db.Genres)
                 {
-                    //Add Error message here
-                    return View();
+                    if (genre.GenreID == tID)
+                    {
+                        //Add Error message here
+                        return View();
+                    }
                 }
+
+                //Genre does not exist, proceed with the addition to DB
+                Genre nuGenre = new Genre() { GenreID = tID, GenreName = _model.GenreName };
+
+                //Add it to the DB
+                db.Genres.Add(nuGenre);
+
+                //Add a test movie
+                //Movie tMovie = new Movie() { MovieGenre = nuGenre, MovieID = 0, MovieName = "TestMovie", ReleaseDate = System.DateTime.Today };
+                //nuGenre.Movies.Add(tMovie);
+                //db.Movies.Add(tMovie);
+
+                //Commit
+                db.SaveChanges();
+
+                return RedirectToAction("Index", "MovieLib");
             }
 
-            //Genre does not exist, proceed with the addition to DB
-            Genre nuGenre = new Genre() { GenreID = tID, GenreName = _model.GenreName};
-
-            //Add it to the DB
-            db.Genres.Add(nuGenre);
-            //Commit
-            db.SaveChanges();
-
-            return RedirectToAction("Index", "MovieLib");
+            return RedirectToAction("AddGenre", "MovieLib");
         }
 
         [HttpGet]
@@ -73,24 +85,89 @@ namespace TheMovieDB.Controllers
             {
                 IdentityDBContext db = new IdentityDBContext();
 
+                var myModel = db.Genres.Include(m => m.Movies).SingleOrDefault(m => m.GenreID == _genreID);
+
+                List<Movie> movieList = myModel.Movies.ToList();
+
                 foreach(Genre genre in db.Genres)
                 {
-                    if(genre.GenreID == _genreID)
+                    if(genre.GenreID == _genreID && movieList.Count == 0)
                     {
                         //Delete the Genre
                         db.Genres.Remove(genre);
                         break;
                     }
-                    //else if(genre.GenreID == _genreID && genre.Movies.Count > 0)
-                    //{
-                    //    //add Errors here
-                    //    break;
-                    //}
+                    else if (genre.GenreID == _genreID && movieList.Count > 0)
+                    {
+                        //add Errors here
+                        Console.WriteLine("Error when deleting Genre");
+                        break;
+                    }
                 }
 
                 //save changes
                 db.SaveChanges();
             }
+            return RedirectToAction("Index", "MovieLib");
+        }
+
+        [HttpGet]
+        public ActionResult ViewMovies(int _genreID)
+        {
+            IdentityDBContext dbContext = new IdentityDBContext();
+
+            var context = dbContext.Genres.Include(m => m.Movies).SingleOrDefault(m => m.GenreID == _genreID);
+
+            ViewMovieModel model = new ViewMovieModel() { MovieGenreID = _genreID, MovieList = context.Movies.ToList() };
+
+            return View(model);
+        }
+
+        [HttpGet]
+        public ActionResult AddMovie(int _genreID)
+        {
+            if(Request.IsAuthenticated)
+            {
+                IdentityDBContext db = new IdentityDBContext();
+
+                Genre tGenre = db.Genres.Find(_genreID);
+
+                if(tGenre != null)
+                {
+                    AddMovieModel tModel = new AddMovieModel() { MovieGenreID = tGenre.GenreID, MovieGenreName = tGenre.GenreName };
+
+                    return View(tModel);
+                }
+
+                return RedirectToAction("Index", "MovieLib");
+            }
+            return RedirectToAction("LoginUser", "Login");
+        }
+
+        [HttpPost]
+        public ActionResult AddMovie(AddMovieModel _model)
+        {
+            if(ModelState.IsValid)
+            {
+                IdentityDBContext db = new IdentityDBContext();
+
+                Genre tGenre = db.Genres.Find(_model.MovieGenreID);
+
+                Movie tMovie = new Movie() { MovieID = GetMovieID(_model.MovieName), MovieName = _model.MovieName, ReleaseDate = _model.ReleaseDate, MovieGenre = tGenre, MovieGenreID = tGenre.GenreID };
+
+                //Add it to the Genre list
+                var genreContext = db.Genres.Include(m => m.Movies).SingleOrDefault(m => m.GenreID == _model.MovieGenreID);
+                genreContext.Movies.ToList().Add(tMovie);
+
+                //Now, add the Movie to the DB
+                db.Movies.Add(tMovie);
+
+                //Save changes in the DB
+                db.SaveChanges();
+
+                return RedirectToAction("ViewMovies", "MovieLib", new { _genreID = _model.MovieGenreID });
+            }
+
             return RedirectToAction("Index", "MovieLib");
         }
